+++ ShyamGui/Source/MeasurementData.h
@@ -1184,7 +1184,7 @@ namespace MeasurementData
         };
 
         line ("SourceDesc=Atomik_Data_Text");
-        line ("Version='beta_v1.3.0'");
+        line ("Version='v1.3.0'");
         line ("Author='Atomik Prediction Software'");
         line ("");
         line ("Frequency_Hz=" + juce::String (hz));